add error handling for cart service methods

+++ src/services/cartService.ts
@@ -1,8 +1,9 @@
 import prisma from "../config/prisma";
+import { ResponseError } from "../errors/responseError";
 
 export class CartService {
  
-  static async getCart(customerId: string) {
+  static async getCustomerCart(customerId: string) {
     const carts = await prisma.cart.findMany({
       where: { customerId: parseInt(customerId) },
       include: {
@@ -18,6 +19,19 @@ export class CartService {
   }
 
   static async removeCart(req: { customerId: number, productId: number }) {
+    const getCart  = await prisma.cart.findUnique({
+      where: {
+        customerId_productId: {
+          customerId: req.customerId,
+          productId: req.productId
+        }
+      },
+    });
+
+    if (!getCart) {
+      throw new ResponseError(404, "Item not found in cart");
+    }
+
     const cart = await prisma.cart.delete({
       where: {
         customerId_productId: {
@@ -30,6 +44,16 @@ export class CartService {
   }
 
   static async removeAllCart(req: { customerId: number }) {
+    const getCart = await prisma.cart.findFirst({
+      where: {
+        customerId: req.customerId
+      }
+    });
+
+    if (!getCart) {
+      throw new ResponseError(404, "Item not found in cart");
+    }
+
     const cart = await prisma.cart.deleteMany({
       where: {
         customerId: req.customerId
